% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_index.R
\name{combination_index}
\alias{combination_index}
\title{compute combination index}
\usage{
combination_index(MACSQuant, ...)
}
\arguments{
\item{MACSQuant}{object of class MACSQuant resulting of the function
load_maxQuant().
Contains the original data table}

\item{...}{params for lattice cloud namely z and x for parameter screen
default for z and x are c(-110,-70) and argument xlab and ylab}
}
\value{
Several plots and combination index
}
\description{
This function allows the user to compute combination index
on the drug combinations. This function also generates intermediary
plots and tables.
}
\examples{
filepath <- system.file("extdata", "drugs.Rdata",
    package = "MACSQuantifyR")
load(filepath)
combination_index(MACSQuant)
}
\references{
Chou, T. C. (2006). Theoretical basis, experimental design, and
computerized simulation of synergism and antagonism in drug
combination studies. \emph{Pharmacological reviews}, 58(3), 621-681.
}
