% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.r
\name{generate_report}
\alias{generate_report}
\title{Word report generation}
\usage{
generate_report(MACSQuant)
}
\arguments{
\item{MACSQuant}{object of class MACSQuant resulting of the function
load_maxQuant().
Contains the original data table}
}
\value{
A formatted report
}
\description{
Allows the user to generate a report at every
steps of the analysis

Also generates a complete report with the graphics
}
\examples{
drugs_R_image <- system.file("extdata",
    "drugs.RDS",
    package = "MACSQuantifyR")
MACSQuant <- readRDS(drugs_R_image)
generate_report(MACSQuant)
}
