% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_MACSQuant.R
\name{load_MACSQuant}
\alias{load_MACSQuant}
\title{xls file from maxQuant 96-well-plate device}
\usage{
load_MACSQuant(filepath,sheet_name)
}
\arguments{
\item{filepath}{path of the excel file}

\item{sheet_name}{Name of the sheet to load
(optional, first sheet is default)}
}
\value{
An object called MACSQuant of class MACSQuant containing
variable my_data that corresponds to the data of the excel file in R
}
\description{
xls file from maxQuant 96-well-plate device
}
\examples{
filepath=system.file('extdata', 'SingleDrugs.xlsx',
package = 'MACSQuantifyR')
# load_MACSQuant(filepath)
}
