% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MQ.R
\docType{class}
\name{MACSQuant}
\alias{MACSQuant}
\alias{MACSQuant-class}
\title{The MACSQuant Class}
\description{
MACSQuant object contains all the data and the user parameters
}
\section{Slots}{

\describe{
\item{\code{my_data}}{Contains the raw data}

\item{\code{my_data_sorted}}{Contains the sorted data
(according to replicates order)}

\item{\code{my_replicates_sorted}}{Contains the sorted matrix containing
replicates names}

\item{\code{experiment_name}}{Contains the experiment name
given by the user}

\item{\code{statistics}}{Contains the statistics for each condition}

\item{\code{combination.index.df}}{data.frame that contains the outputs of
the combination_index function}

\item{\code{param.experiment}}{list that contains the parameters relative
to the experiment (i.e, number of replicates,
number of conditions...)}

\item{\code{param.output}}{List that contains the parameters relative
to the report generation and outputs (i.e, save intermediary
results, generate full report..)}
}}

