% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_function_toolbox.R
\name{rData}
\alias{rData}
\title{accessor function to access raw data}
\usage{
rData(MACSQuant)
}
\arguments{
\item{MACSQuant}{object of class MACSQuant resulting of the function
load_maxQuant().
Contains the original data table}
}
\value{
the raw data table
}
\description{
This function allows the user to access raw data table
}
\examples{
filepath <- system.file("extdata", "SingleDrugs.xlsx",
    package = "MACSQuantifyR")
#MACSQuant = load_MACSQuant(filepath)

#rData(MACSQuant)
}
