% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_function_toolbox.R
\name{setPath}
\alias{setPath}
\title{sets custom output path}
\usage{
setPath(MACSQuant,path=NULL)
}
\arguments{
\item{MACSQuant}{object of class MACSQuant resulting of the function
load_maxQuant().
Contains the original data table}

\item{path}{user defined path, default is tempd}
}
\value{
object of class MACSQuant with updated path
}
\description{
This function allows the user to set the output directory path
}
\examples{
filepath <- system.file("extdata", "SingleDrugs.xlsx",
    package = "MACSQuantifyR")
#MACSQuant = load_MACSQuant(filepath)
user_path="."
#MACSQuant = setPath(MACSQuant,path=user_path)
}
